[feat] implement thread tiling runtime dispatch; change index type from uncessary size_t to int

+++ src/gemm/sgemm_naive.cu
@@ -3,28 +3,28 @@
 
 #include <cuda_runtime.h>
 
-#include <cstddef>
-
 namespace gemm {
 
 namespace {
 
-template <typename T, size_t BLOCK_SIZE_X, size_t BLOCK_SIZE_Y>
-__global__ void sgemm_naive(size_t m, size_t n, size_t k, T const* A,
+template <typename T, int BLOCK_SIZE_X, int BLOCK_SIZE_Y>
+__global__ void sgemm_naive(int m, int n, int k, T const* A,
                             T const* B, T const* C, T* D, T alpha, T beta) {
-  size_t const row{blockIdx.y * BLOCK_SIZE_Y + threadIdx.y};
-  size_t const col{blockIdx.x * BLOCK_SIZE_X + threadIdx.x};
+  int const row{static_cast<int>(blockIdx.y) * BLOCK_SIZE_Y +
+                static_cast<int>(threadIdx.y)};
+  int const col{static_cast<int>(blockIdx.x) * BLOCK_SIZE_X +
+                static_cast<int>(threadIdx.x)};
 
   if (row >= m || col >= n) {
     return;
   }
 
   T sum{};
-  for (size_t inner{}; inner < k; ++inner) {
+  for (int inner{}; inner < k; ++inner) {
     sum += A[row * k + inner] * B[inner * n + col];
   }
 
-  size_t const idx{row * n + col};
+  int const idx{row * n + col};
   T const previous{(beta != T{} && C != nullptr) ? C[idx] : T{}};
   D[idx] = alpha * sum + beta * previous;
 }
@@ -34,8 +34,8 @@ __global__ void sgemm_naive(size_t m, size_t n, size_t k, T const* A,
 void launch_sgemm_naive(const SgemmProblem &problem, const float *a,
                         const float *b, const float *c, float *d,
                         cudaStream_t stream) {
-  constexpr size_t BLOCK_SIZE_X{16};
-  constexpr size_t BLOCK_SIZE_Y{16};
+  constexpr int BLOCK_SIZE_X{16};
+  constexpr int BLOCK_SIZE_Y{16};
   dim3 const block_dim{static_cast<unsigned int>(BLOCK_SIZE_X),
                        static_cast<unsigned int>(BLOCK_SIZE_Y), 1};
   dim3 const grid_dim{
@@ -44,10 +44,9 @@ void launch_sgemm_naive(const SgemmProblem &problem, const float *a,
       1};
 
   sgemm_naive<float, BLOCK_SIZE_X, BLOCK_SIZE_Y>
-      <<<grid_dim, block_dim, 0, stream>>>(
-          static_cast<size_t>(problem.m), static_cast<size_t>(problem.n),
-          static_cast<size_t>(problem.k), a, b, c, d, problem.alpha,
-          problem.beta);
+      <<<grid_dim, block_dim, 0, stream>>>(problem.m, problem.n, problem.k, a,
+                                           b, c, d, problem.alpha,
+                                           problem.beta);
   GEMM_CUDA_CHECK(cudaGetLastError());
 }
