+++ src/gemm/sgemm_smem.cu
@@ -3,32 +3,32 @@
 
 #include <cuda_runtime.h>
 
-#include <cstddef>
-
 namespace gemm {
 
 namespace {
 
-template <typename T, size_t BLOCK_TILE_SIZE_X, size_t BLOCK_TILE_SIZE_Y,
-          size_t BLOCK_TILE_SIZE_K>
-__global__ void sgemm_smem(size_t m, size_t n, size_t k, T const* A,
+template <typename T, int BLOCK_TILE_SIZE_X, int BLOCK_TILE_SIZE_Y,
+          int BLOCK_TILE_SIZE_K>
+__global__ void sgemm_smem(int m, int n, int k, T const* A,
                            T const* B, T const* C, T* D, T alpha, T beta) {
   __shared__ T A_thread_block_tile[BLOCK_TILE_SIZE_Y][BLOCK_TILE_SIZE_K];
   __shared__ T B_thread_block_tile[BLOCK_TILE_SIZE_K][BLOCK_TILE_SIZE_X];
 
-  size_t const row{blockIdx.y * BLOCK_TILE_SIZE_Y + threadIdx.y};
-  size_t const col{blockIdx.x * BLOCK_TILE_SIZE_X + threadIdx.x};
+  int const row{static_cast<int>(blockIdx.y) * BLOCK_TILE_SIZE_Y +
+                static_cast<int>(threadIdx.y)};
+  int const col{static_cast<int>(blockIdx.x) * BLOCK_TILE_SIZE_X +
+                static_cast<int>(threadIdx.x)};
 
   T sum{};
 
-  size_t const num_thread_block_tiles{(k + BLOCK_TILE_SIZE_K - 1) /
-                                      BLOCK_TILE_SIZE_K};
-  for (size_t thread_block_tile_idx{}; thread_block_tile_idx <
-                                      num_thread_block_tiles;
+  int const num_thread_block_tiles{(k + BLOCK_TILE_SIZE_K - 1) /
+                                   BLOCK_TILE_SIZE_K};
+  for (int thread_block_tile_idx{}; thread_block_tile_idx <
+                                   num_thread_block_tiles;
        ++thread_block_tile_idx) {
-    size_t const A_row_idx{row};
-    size_t const A_col_idx{thread_block_tile_idx * BLOCK_TILE_SIZE_K +
-                           threadIdx.x};
+    int const A_row_idx{row};
+    int const A_col_idx{thread_block_tile_idx * BLOCK_TILE_SIZE_K +
+                        static_cast<int>(threadIdx.x)};
     if (A_row_idx < m && A_col_idx < k) {
       A_thread_block_tile[threadIdx.y][threadIdx.x] =
           A[A_row_idx * k + A_col_idx];
@@ -36,9 +36,9 @@ __global__ void sgemm_smem(size_t m, size_t n, size_t k, T const* A,
       A_thread_block_tile[threadIdx.y][threadIdx.x] = T{};
     }
 
-    size_t const B_row_idx{thread_block_tile_idx * BLOCK_TILE_SIZE_K +
-                           threadIdx.y};
-    size_t const B_col_idx{col};
+    int const B_row_idx{thread_block_tile_idx * BLOCK_TILE_SIZE_K +
+                        static_cast<int>(threadIdx.y)};
+    int const B_col_idx{col};
     if (B_row_idx < k && B_col_idx < n) {
       B_thread_block_tile[threadIdx.y][threadIdx.x] =
           B[B_row_idx * n + B_col_idx];
@@ -47,7 +47,7 @@ __global__ void sgemm_smem(size_t m, size_t n, size_t k, T const* A,
     }
     __syncthreads();
 
-    for (size_t k_i{}; k_i < BLOCK_TILE_SIZE_K; ++k_i) {
+    for (int k_i{}; k_i < BLOCK_TILE_SIZE_K; ++k_i) {
       sum += A_thread_block_tile[threadIdx.y][k_i] *
              B_thread_block_tile[k_i][threadIdx.x];
     }
@@ -56,7 +56,7 @@ __global__ void sgemm_smem(size_t m, size_t n, size_t k, T const* A,
   }
 
   if (row < m && col < n) {
-    size_t const idx{row * n + col};
+    int const idx{row * n + col};
     T const previous{(beta != T{} && C != nullptr) ? C[idx] : T{}};
     D[idx] = alpha * sum + beta * previous;
   }
@@ -67,9 +67,9 @@ __global__ void sgemm_smem(size_t m, size_t n, size_t k, T const* A,
 void launch_sgemm_smem(const SgemmProblem &problem, const float *a,
                        const float *b, const float *c, float *d,
                        cudaStream_t stream) {
-  constexpr size_t BLOCK_TILE_SIZE_X{16};
-  constexpr size_t BLOCK_TILE_SIZE_Y{16};
-  constexpr size_t BLOCK_TILE_SIZE_K{16};
+  constexpr int BLOCK_TILE_SIZE_X{16};
+  constexpr int BLOCK_TILE_SIZE_Y{16};
+  constexpr int BLOCK_TILE_SIZE_K{16};
   dim3 const block_dim{static_cast<unsigned int>(BLOCK_TILE_SIZE_X),
                        static_cast<unsigned int>(BLOCK_TILE_SIZE_Y), 1};
   dim3 const grid_dim{
@@ -80,10 +80,9 @@ void launch_sgemm_smem(const SgemmProblem &problem, const float *a,
       1};
 
   sgemm_smem<float, BLOCK_TILE_SIZE_X, BLOCK_TILE_SIZE_Y, BLOCK_TILE_SIZE_K>
-      <<<grid_dim, block_dim, 0, stream>>>(
-          static_cast<size_t>(problem.m), static_cast<size_t>(problem.n),
-          static_cast<size_t>(problem.k), a, b, c, d, problem.alpha,
-          problem.beta);
+      <<<grid_dim, block_dim, 0, stream>>>(problem.m, problem.n, problem.k, a,
+                                           b, c, d, problem.alpha,
+                                           problem.beta);
   GEMM_CUDA_CHECK(cudaGetLastError());
 }
